requirements.txt
accelerate==1.0.1
addict==2.4.0
aiohttp==3.8.1
aiosignal==1.2.0
argcomplete==2.0.0
async-timeout==4.0.2
attrs==19.3.0
bleach==6.1.0
boto==2.49.0
cachetools==5.0.0
certifi==2019.11.28
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.8
clip==1.0
colorama==0.4.3
command-not-found==0.3
contourpy==1.1.1
crcmod==1.7
cryptography==36.0.2
cycler==0.12.1
Cython==0.29.14
dbus-python==1.2.16
defusedxml==0.7.1
distlib==0.4.0
distro-info==0.23+ubuntu1.1
docker-pycreds==0.4.0
docutils==0.16
fail2ban==0.11.1
fasteners==0.17.3
filelock==3.16.1
fonttools==4.57.0
frozenlist==1.3.0
fsspec==2025.3.0
ftfy==6.2.3
gcs-oauth2-boto-plugin==3.0
gitdb==4.0.9
GitPython==3.1.27
google-apitools==0.5.32
google-auth==2.6.2
google-reauth==0.1.1
hf-xet==1.2.0
httplib2==0.20.4
huggingface-hub==0.36.0
idna==2.8
importlib-metadata==1.5.0
importlib-resources==6.4.5
iotop==0.6
jinja2==3.1.6
jsonschema==3.2.0
kaggle==1.7.4.5
kiwisolver==1.4.7
language-selector==0.1
MarkupSafe==2.1.5
matplotlib==3.7.5
monotonic==1.6
more-itertools==4.2.0
mpmath==1.3.0
multidict==6.0.2
netifaces==0.10.4
networkx==3.1
numpy==1.21.5
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
nvsm==23.9.2
nvsm-dshm==23.9.2
nvsm-health==23.9.2
nvsm-sdk==1.0.0
oauth2client==4.1.3
opencv-python==4.12.0.88
packaging==20.3
paho-mqtt==1.5.0
pandas==2.0.3
pathtools==0.1.2
pesq==0.0.4
pexpect==4.6.0
pillow==10.4.0
platformdirs==4.3.6
promise==2.3
protobuf==3.20.1
psutil==5.5.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.7
pycparser==2.21
PyGObject==3.36.0
PyJWT==1.7.1
pyOpenSSL==22.0.0
pyparsing==3.0.7
pyrsistent==0.15.5
pystoi==0.4.1
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2025.2
pyu2f==0.1.5
PyYAML==5.3.1
rapidfuzz==3.9.7
regex==2024.11.6
requests==2.32.4
requests-unixsocket==0.2.0
retry-decorator==1.1.1
roman==2.0.0
rsa==4.7.2
safetensors==0.5.3
scipy==1.10.0
screen-resolution-extra==0.0.0
sentry-sdk==1.9.7
setproctitle==1.3.2
shortuuid==1.0.9
six==1.14.0
smmap==5.0.0
sos==4.7.2
speechmos==0.0.1.1
supervision==0.25.1
sympy==1.13.3
text-unidecode==1.3
timm==1.0.22
tokenizers==0.20.3
tomli==2.3.0
torch==2.4.1
torchvision==0.19.1
tqdm==4.67.1
transformers==4.46.3
triton==3.0.0
typing-extensions==4.13.2
tzdata==2025.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
UNKNOWN==0.0.0
urllib3==1.26.12
uv==0.8.3
virtualenv==20.32.0
wandb==0.13.2
wcwidth==0.2.14
webencodings==0.5.1
xkit==0.0.0
yapf==0.43.0
yarl==1.7.2
zipp==3.20.2
